fraud-detection-system/requirements.txt
# Kafka
kafka-python==2.0.2
confluent-kafka==2.1.1

# Data Processing
pandas==2.0.3
numpy==1.24.4

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
imbalanced-learn==0.11.0
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2

# Data Generation
Faker==18.13.0

# Stream Processing
pyflink==1.17.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20

# API and Services
Flask==2.3.3
pyyaml==6.0.1
requests==2.31.0

# Testing
pytest==7.4.0

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2